84-harden-controller-firmware
Why:

- Avoid sign-compare errors
- Avoid sprintf buffer truncation
- Remove unused variables
- Catch unintended switch cases
- Proper member construction ordering

This change addresses the need by:

- Enabling -Wall on debug builds
- Resolving warnings
- Adding safe integer library

Closes #84

## Summary by Sourcery

Enable strict compiler warnings, resolve resulting compile warnings, and harden code by integrating a safe integer library and cleaning up unsafe patterns.

New Features:
- Add Safe integer library for safer PWM channel handling.

Enhancements:
- Bump custom firmware version to 1.15.2.
- Enable -Wall on debug builds and propagate build flags/unflags across all environments.
- Increase buffer sizes for time and date formatting to prevent truncation.
- Convert loop indices from signed int to size_t or uint32_t to avoid sign-compare warnings.
- Remove unused variables and redundant locals from various modules.
- Add default cases in switch statements to catch illegal states.
- Reorder class member declarations to ensure proper construction ordering.

Build:
- Update platformio.ini to enable -Wall, remove -Wno-sign-compare suppression, and include the Safe library dependency.

Documentation:
- Include license file for the Safe library.

+++ src/peripheral/peripherals/modbus/modbus_client_input.cpp
@@ -48,7 +48,7 @@ ModbusClientInput::ModbusClientInput(const JsonObjectConst& parameters)
     return;
   }
   inputs_.resize(inputs_json.size());
-  for (int i = 0; i < inputs_json.size(); i++) {
+  for (size_t i = 0; i < inputs_json.size(); i++) {
     Input& input = inputs_[i];
     JsonObjectConst input_json = inputs_json[i];
